pyproject.toml
[project]
name = "sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ansible-core~=2.17.0",
    "azure-ai-evaluation>=1.12.0",
    "azure-ai-projects~=1.0.0b11",
    "azure-identity~=1.20.0",
    "ipykernel>=7.1.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
]
